+++ packages/utils/abi/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @0xsequence/abi
 
+## 3.0.0-beta.12
+
+### Patch Changes
+
+- Beta release with dapp connector fixes
+
+## 3.0.0-beta.11
+
+### Patch Changes
+
+- 3.0.0 beta
+
 ## 3.0.0-beta.10
 
 ### Patch Changes